useeffect fix to loader
