+++ client/src/pages/Profile.tsx
@@ -1,0 +1,59 @@
+import "../styles/profile.css";
+import { useLoaderData } from "react-router-dom";
+
+export default function Profile() {
+  const userData = useLoaderData() as UserData;
+
+  return (
+    <div className="profile">
+      {userData && (
+        <>
+          <img
+            src={
+              userData.profile_picture ||
+              "https://www.vhv.rs/dpng/d/138-1383989_default-svg-icon-free-avatar-png-transparent-png.png"
+            }
+            alt="pdeprofil"
+          />
+          <div className="username-followers">
+            <h1>{userData.username}</h1>
+            <p>Following: {userData.following}</p>
+            <p>Followers: {userData.followers}</p>
+          </div>
+          <p>Citations</p>
+          <div className="biography">
+            <h3>Biographie</h3>
+            <p>{userData.bio || "Aucune biographie renseignée."}</p>
+          </div>
+          <div className="links">
+            <h3>Liens</h3>
+            <ul className="list-none">
+              {userData.portfolio && (
+                <li>
+                  <a
+                    href={userData.portfolio}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                  >
+                    Portfolio
+                  </a>
+                </li>
+              )}
+              {userData.website && (
+                <li>
+                  <a
+                    href={userData.website}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                  >
+                    Website
+                  </a>
+                </li>
+              )}
+            </ul>
+          </div>
+        </>
+      )}
+    </div>
+  );
+}
