+++ client/src/services/requests.ts
@@ -13,4 +13,14 @@ const getAllArtwork = async () => {
   }
 };
 
-export { getAllArtwork };
+const getUserById = async (id: number) => {
+  try {
+    const response = await axios.get(`${baseUrl}/api/user/${id}`);
+    return response.data;
+  } catch (error) {
+    console.error(error);
+    throw new Error("Failed to fetch users");
+  }
+};
+
+export { getAllArtwork, getUserById };
