+++ client/src/types/index.d.ts
@@ -10,3 +10,16 @@ interface Artwork {
 interface Artworks {
   artworks: Artwork[];
 }
+
+interface UserData {
+  id: number;
+  username: string;
+  email: string;
+  password: string;
+  profile_picture: string | null;
+  following: number;
+  followers: number;
+  bio: string | null;
+  portfolio: string | null;
+  website: string | null;
+}
